Avoid int

+++ libsrl/include/srl/srl_sched_wait.h
@@ -30,7 +30,7 @@ DECLARE_WAIT(cpu, gt, >)
 
 #undef DECLARE_WAIT
 
-typedef int srl_callback_t( uint32_t val );
+typedef int8_t srl_callback_t( uint32_t val );
 
 void srl_sched_wait_priv( srl_callback_t *cb, uint32_t val );
